+++ django_admin_logs/settings.py
@@ -10,3 +10,9 @@ DJANGO_ADMIN_LOGS_ENABLED = getattr(settings, "DJANGO_ADMIN_LOGS_ENABLED", True)
 # Determines whether admin logs are deletable (defaults to False).
 # If enabled, non supers users will still require the delete_logentry permission.
 DJANGO_ADMIN_LOGS_DELETABLE = getattr(settings, "DJANGO_ADMIN_LOGS_DELETABLE", False)
+
+# Determines whether to ignore (not log) CHANGE actions where no changes were made
+# (defaults to False).
+DJANGO_ADMIN_LOGS_IGNORE_UNCHANGED = getattr(
+    settings, "DJANGO_ADMIN_LOGS_IGNORE_UNCHANGED", False
+)